+++ MacForge/MacForge/Views/Web/MF_searchView.m
@@ -29,7 +29,7 @@ extern AppDelegate *myDelegate;
 
 - (void)adjustColumnWidth {
     int multiplier = 4;
-    if (MF_extra.sharedInstance.macOS >= 16) multiplier = 25;
+//    if (MF_extra.sharedInstance.macOS >= 16) multiplier = 25;
     NSUInteger pad = self.frame.size.width - (multiplier * columns);
     for (NSTableColumn* c in _tv.tableColumns)
         [c setWidth:pad/columns];
@@ -67,7 +67,7 @@ extern AppDelegate *myDelegate;
         smallArray = NSMutableArray.new;
         
         NSRect theFrame = self.frame;
-        if (MF_extra.sharedInstance.macOS >= 16) theFrame.size.height += 38;
+//        if (MF_extra.sharedInstance.macOS >= 16) theFrame.size.height += 38;
         
         // Create a table view
         _tv = [[NSTableView alloc] initWithFrame:theFrame];